Monster can also be controlled by 2nd Player
In this version, the game comes for 2 players. Now you can also control the monster (with a,s,w,d keys). Our Hero needs to catch the monster to get the point, Also if the Monster goes out of the boundary our Hero gets a point. But if our Hero goes out of the boundary the  Monster will get a point.